requirements.txt
python-dotenv>=1.0.1
pydantic>=2.7.0
langchain>=0.2.10
langchain-core>=0.2.28
langchain-google-genai>=2.0.0
langchain-tavily
langgraph